+++ src/java/com/android/internal/telephony/SubscriptionInfoUpdater.java
@@ -395,7 +395,7 @@ public class SubscriptionInfoUpdater extends Handler {
                 number.put(SubscriptionManager.NUMBER, msisdn);
                 contentResolver.update(SubscriptionManager.CONTENT_URI, number,
                         SubscriptionManager.UNIQUE_KEY_SUBSCRIPTION_ID + "="
-                        + Long.toString(subId), null);
+                        + Integer.toString(subId), null);
             }
 
             SubscriptionInfo subInfo = mSubscriptionManager.getActiveSubscriptionInfo(subId);
@@ -414,7 +414,7 @@ public class SubscriptionInfoUpdater extends Handler {
                 logd("sim name = " + nameToSet);
                 contentResolver.update(SubscriptionManager.CONTENT_URI, name,
                         SubscriptionManager.UNIQUE_KEY_SUBSCRIPTION_ID
-                        + "=" + Long.toString(subId), null);
+                        + "=" + Integer.toString(subId), null);
             }
 
         } else {
