+++ src/java/com/android/internal/telephony/gsm/GsmServiceStateTracker.java
@@ -2050,7 +2050,7 @@ final class GsmServiceStateTracker extends ServiceStateTracker {
 
         switch (notifyType) {
         case PS_ENABLED:
-            long dataSubId = SubscriptionManager.getDefaultDataSubId();
+            int dataSubId = SubscriptionManager.getDefaultDataSubId();
             if (dataSubId != mPhone.getSubId()) {
                 return;
             }
